Fix transposition table usage in searcher
Previously, the searcher was using the transposition table entry for a
given zobrist hash if the depth was `>=` the current depth. However, since
in the context of our searcher, we use decreasing depths (a lower depth
indicates that more depths have been searched) we want to use the
entries when the depth is `<=` the current depth.

I noticed this because the bot was drawing in cases where it had an
obvious checkmate.